Manual merge of main into CETERRA-42-Implement-Game-Behaviour

+++ Assets/Code/Scripts/Domains/Board/Field/CatastropheFieldBehaviour.cs
@@ -42,7 +42,7 @@ public class CatastropheFieldBehaviour : FieldBehaviour {
                     }
                 }
             }
-            BoardContext.Instance.UpdateEnvironmentStat((uint)environmentEffect);
+            BoardContext.Instance.UpdateEnvironmentStat(environmentEffect);
         }
         CompleteFieldInvocation();
     }
